requirements.txt
tqdm
wandb
torchvision
gdown